Include the modules inside the main file to have access to them
